Correcao da atividade
