--- challenges/01-reverse/src/StringReverser.java
@@ -1,31 +1,0 @@
-import java.util.Scanner;
-
-public class StringReverser {
-
-    public static void main(String[] args) {
-        
-        try(Scanner scanner = new Scanner(System.in)){
-
-            System.out.print("Digite uma string: ");
-            String input = scanner.nextLine();
-
-            System.out.printf("String invertida: %s\n", reverse(input));
-       }
-    }
-
-    public String reverse(String inputString) {
-
-        if (inputString == "") {
-            throw new IllegalArgumentException("A string não pode ser nula.");
-        }
-
-        String reversedString = new String();
-  
-        for (int i = inputString.length(); i > 0; i--) {
-            reversedString += inputString.charAt(i);
-        }
-        
-        return reversedString;
- 
-    }
-}
