+++ challenges/01-reverse/src/guilherme-portas allan-samuel gabriela-soares/StringReverser.java
@@ -1,0 +1,31 @@
+import java.util.Scanner;
+
+public class StringReverser {
+
+    public static void main(String[] args) {
+        
+        try(Scanner scanner = new Scanner(System.in)){
+
+            System.out.print("Digite uma string: ");
+            String input = scanner.nextLine();
+
+            System.out.printf("String invertida: %s\n", StringReverser.reverse(input));
+       }
+    }
+
+    public static String reverse(String inputString) {
+
+        if (inputString.strip().equals("")) {
+            throw new IllegalArgumentException("A string não pode ser nula.");
+        }
+
+        String reversedString = new String();
+  
+        for (int i = inputString.length() - 1; i >= 0; i--) {
+            reversedString += inputString.charAt(i);
+        }
+        
+        return reversedString;
+ 
+    }
+}
